refactor: move audio batch structures to lib/audio where they belong
- Move audio_batch_info_t typedef from lib/util/audio.h to lib/audio/audio.h
- Move audio_parse_batch_header() implementation from lib/util/audio.c to lib/audio/audio.c
- Move audio_validate_batch_params() and audio_is_supported_sample_rate() to lib/audio/audio.c
- Update includes in packet_parsing.h to use audio/audio.h instead of util/audio.h
- Remove redundant packet_parse_audio_batch_header() from packet_parsing.c
- Create backwards-compatible inline wrapper in packet_parsing.h
- Delete lib/util/audio.h and lib/util/audio.c files
- Update lib/util/util.h to remove audio.h include
- Update src/server/protocol.c to remove util/audio.h include

Audio batch packet parsing utilities now correctly reside in the audio module
where they logically belong, eliminating a misplaced header in util.
